+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Bumped dbatools to 2.0.0 for the integration tests.
+
 ## [16.3.0] - 2023-04-26
 
 ### Remove