+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- SqlServerDsc
+  - Temporary disable integration tests for dbatools.
+
 ## [16.3.0] - 2023-04-26
 
 ### Remove